Give the callback kinds and search directions single homes
`kCallback*` was the most scattered name in the crate: 92 modules
declared it, one line each, for four values. It settles in
eval/typval.rs, where `Callback` itself is built and read.

`FORWARD`/`BACKWARD` and the two `*_FILE` variants that only search.rs
had settle there with them, 49 modules lighter. Two rewritten call sites
compared `type_0 as c_int` against the constant; with one definition the
cast goes.

+++ crates/nvim/src/nvim/channel.rs
@@ -16,7 +16,7 @@
 use core::ffi::{CStr, c_char, c_int, c_void};
 use core::{mem, ptr};
 
-use crate::src::nvim::eval::typval::{callback_free, tv_dict_free};
+use crate::src::nvim::eval::typval::{callback_free, kCallbackNone, tv_dict_free};
 use crate::src::nvim::event::libuv_proc::libuv_proc_init;
 use crate::src::nvim::event::r#loop::one_arg_event;
 use crate::src::nvim::event::multiqueue::{
@@ -799,7 +799,7 @@ unsafe extern "C" fn channel_proc_exit_cb(_proc: *mut Proc, status: c_int, data:
         }
     }
     let exited = status >= 0;
-    if exited && (*chan).on_exit.type_0 as c_int != kCallbackNone {
+    if exited && (*chan).on_exit.type_0 != kCallbackNone {
         schedule_channel_event(chan);
     }
     if exited {
